% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData4metdna.R
\name{checkData4metdna}
\alias{checkData4metdna}
\title{checkData4metdna}
\usage{
checkData4metdna(path = ".")
}
\arguments{
\item{path}{Absolute directory}
}
\value{
code
}
\description{
checkData4metdna
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
